+++ FixClient/SessionForm.Designer.cs
@@ -43,52 +43,73 @@
             this.OK = new System.Windows.Forms.Button();
             this.Cancel = new System.Windows.Forms.Button();
             this._gridPlaceHolder = new System.Windows.Forms.Panel();
+            this._buttonPanel = new System.Windows.Forms.FlowLayoutPanel();
+            this._buttonPanel.SuspendLayout();
             this.SuspendLayout();
             // 
             // OK
             // 
-            this.OK.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
+            this.OK.AutoSize = true;
             this.OK.DialogResult = System.Windows.Forms.DialogResult.OK;
-            this.OK.Location = new System.Drawing.Point(297, 712);
+            this.OK.Margin = new System.Windows.Forms.Padding(6, 0, 0, 0);
+            this.OK.MinimumSize = new System.Drawing.Size(88, 27);
             this.OK.Name = "OK";
-            this.OK.Size = new System.Drawing.Size(75, 23);
-            this.OK.TabIndex = 18;
+            this.OK.TabIndex = 0;
             this.OK.Text = "OK";
             // 
             // Cancel
             // 
-            this.Cancel.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
+            this.Cancel.AutoSize = true;
             this.Cancel.DialogResult = System.Windows.Forms.DialogResult.Cancel;
-            this.Cancel.Location = new System.Drawing.Point(378, 712);
+            this.Cancel.Margin = new System.Windows.Forms.Padding(6, 0, 0, 0);
+            this.Cancel.MinimumSize = new System.Drawing.Size(88, 27);
             this.Cancel.Name = "Cancel";
-            this.Cancel.Size = new System.Drawing.Size(75, 23);
-            this.Cancel.TabIndex = 19;
+            this.Cancel.TabIndex = 1;
             this.Cancel.Text = "Cancel";
+            // 
+            // _buttonPanel
+            // 
+            this._buttonPanel.AutoSize = true;
+            this._buttonPanel.AutoSizeMode = System.Windows.Forms.AutoSizeMode.GrowAndShrink;
+            this._buttonPanel.Controls.Add(this.Cancel);
+            this._buttonPanel.Controls.Add(this.OK);
+            this._buttonPanel.Dock = System.Windows.Forms.DockStyle.Bottom;
+            this._buttonPanel.FlowDirection = System.Windows.Forms.FlowDirection.RightToLeft;
+            this._buttonPanel.Name = "_buttonPanel";
+            this._buttonPanel.Padding = new System.Windows.Forms.Padding(0, 12, 0, 0);
+            this._buttonPanel.TabIndex = 1;
+            this._buttonPanel.WrapContents = false;
             // 
             // _gridPlaceHolder
             // 
-            this._gridPlaceHolder.Location = new System.Drawing.Point(12, 12);
+            this._gridPlaceHolder.Dock = System.Windows.Forms.DockStyle.Fill;
             this._gridPlaceHolder.Name = "_gridPlaceHolder";
-            this._gridPlaceHolder.Size = new System.Drawing.Size(441, 681);
-            this._gridPlaceHolder.TabIndex = 35;
+            this._gridPlaceHolder.TabIndex = 0;
             // 
             // SessionForm
             // 
             this.AcceptButton = this.OK;
-            this.ClientSize = new System.Drawing.Size(465, 747);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(7F, 15F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.CancelButton = this.Cancel;
+            this.ClientSize = new System.Drawing.Size(560, 720);
             this.Controls.Add(this._gridPlaceHolder);
-            this.Controls.Add(this.Cancel);
-            this.Controls.Add(this.OK);
-            this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedDialog;
+            this.Controls.Add(this._buttonPanel);
+            this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.Sizable;
             this.Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
             this.MaximizeBox = false;
             this.MinimizeBox = false;
+            this.MinimumSize = new System.Drawing.Size(420, 400);
             this.Name = "SessionForm";
+            this.Padding = new System.Windows.Forms.Padding(12);
             this.ShowInTaskbar = false;
-            this.SizeGripStyle = System.Windows.Forms.SizeGripStyle.Hide;
+            this.SizeGripStyle = System.Windows.Forms.SizeGripStyle.Show;
             this.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent;
             this.Text = "Configuration";
+            this._buttonPanel.ResumeLayout(false);
+            this._buttonPanel.PerformLayout();
             this.ResumeLayout(false);
+            this.PerformLayout();
 
 		}
 
@@ -97,5 +118,6 @@
         private System.Windows.Forms.Button OK;
         private System.Windows.Forms.Button Cancel;
         private System.Windows.Forms.Panel _gridPlaceHolder;
+        private System.Windows.Forms.FlowLayoutPanel _buttonPanel;
 	}
 }